Bug fix + ResourceSet gadget
There is a bug for the --test argument as it doesn't work when it is a bridge gadget. I have fixed this by creating a deep clone object from the InputArgs.

I have renamed the CmdFullString to Cmd in InputArgs to make it less confusing and compatible to what we had before.

I have also added ResourceSet gadget which needs improvment in the future when I find a way to edit binaryformatter without actually creating an object. It uses the plugin we had for resx.

I have also done some code cleanup to remove unnecessary imports.