Another major update
This changes the help section to make it shorter by default but can also show `fullhelp`.

Gadgets now can have their own arguments. For example, some variants and option added to some of the gadgets. This way we can move towards the goal ofresearch as well as pentesting because we don't need to throw out any findings. For example, the Source attribute of ResourceDictionary can be very useful when we have length problem.

In addition to extra arguments that can be passed by users, this release can also support for internal arguments coming through bridge/derived methods - this can be useful to create different payloads when needed based on the requirements.

Gadgets can now override an Init() method if needed to run something before start (default is for reading extra arguments).