Improvements
TypeConfuseDelegate payload has been minified.

A new test app has been added that can be used to identify whether or not an RCE was successful. I will probably add it to the serializer helper test cases in the future too.

The Advanced/SampleBinaryFormatterParser has been updated to support more functions. Some of the methods have been renamed to be be easier to understand.

More methods have been added to the SerializersHelper.

Json.Net has been updated to the latest version so we can use its full power.

A testing arena feature has been added (can be activated by using --runmytest) - this will be useful for testing and debugging as we can just code there rather than changing the main functions of ysoserial.net or without including it as an external reference.