Fixing bugs and readme update
ActivitySurrogateSelectorFromFile did not work without the E.dll and it was missing because of a wrong configuration. In addition to fixing that, ActivitySurrogateSelectorGenerator.cs has been updated so ActivitySurrogateSelectorFromFile will not need the E.dll file in the first place.

The credit section of readme has been updated as it was missing a name.